fix :: 위경도 변환 안되는 오류 수정

+++ src/main/kotlin/com/ondosee/domain/weather/presentation/data/req/QueryTodayWeatherSignificantRequestData.kt
@@ -4,6 +4,6 @@ import org.springframework.web.bind.annotation.RequestParam
 
 data class QueryTodayWeatherSignificantRequestData(
     @RequestParam val location: String,
-    @RequestParam val x: Long,
-    @RequestParam val y: Long
+    @RequestParam val x: Double,
+    @RequestParam val y: Double
 )
