+++ src/main/kotlin/com/ondosee/domain/weather/service/data/req/GetTodayWeatherRequestData.kt
@@ -1,7 +1,7 @@
 package com.ondosee.domain.weather.service.data.req
 
 data class GetTodayWeatherRequestData(
-    val x: Long,
-    val y: Long,
+    val x: Double,
+    val y: Double,
     val location: String
 )
